backend/import_functionality/requirements.txt
anthropic>=0.18.0
pypdf2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0